+++ src/ClientCore/Systems/Camera/CameraSystem.cs
@@ -23,6 +23,7 @@
 
 using Sovereign.EngineCore.Entities;
 using Sovereign.EngineCore.Events;
+using Sovereign.EngineCore.Events.Details;
 using Sovereign.EngineCore.Systems;
 using System;
 using System.Collections.Generic;
@@ -52,13 +53,20 @@ namespace Sovereign.ClientCore.Systems.Camera
         };
 
         public CameraSystem(CameraManager cameraManager, CameraEventHandler eventHandler,
-            IEventLoop eventLoop, EventCommunicator eventCommunicator)
+            IEventLoop eventLoop, EventCommunicator eventCommunicator, 
+            EventDescriptions eventDescriptions)
         {
+            /* Dependency injection. */
             this.cameraManager = cameraManager;
             this.eventHandler = eventHandler;
             this.eventLoop = eventLoop;
             EventCommunicator = eventCommunicator;
 
+            /* Register events. */
+            eventDescriptions.RegisterEvent<EntityEventDetails>(EventId.Client_Camera_Attach);
+            eventDescriptions.RegisterNullEvent(EventId.Client_Camera_Detach);
+
+            /* Register system. */
             eventLoop.RegisterSystem(this);
         }